+++ Tryout/src/main/kotlin/net/milosvasic/try/logger/Main.kt
@@ -47,12 +47,3 @@ fun main(args: Array<String>) {
     }
 
 }
-
-fun getHome(): File {
-    val home = System.getProperty("user.home")
-    val root = File("$home${File.separator}Logger_Logs")
-    if (!root.exists()) {
-        root.mkdirs()
-    }
-    return root
-}