+++ Tryout/src/main/kotlin/net/milosvasic/try/logger/Shared.kt
@@ -1,0 +1,12 @@
+package net.milosvasic.`try`.logger
+
+import java.io.File
+
+fun getHome(): File {
+    val home = System.getProperty("user.home")
+    val root = File("$home${File.separator}Logger_Logs")
+    if (!root.exists()) {
+        root.mkdirs()
+    }
+    return root
+}